Squash merge from ka/sdk_cache_tests.W-7487360 branch
