+++ .circleci/config.yml
@@ -48,7 +48,7 @@ jobs:
     steps:
       - build
       - test
-    
+
   build-node-14:
     executor: node-14
     steps:
