+++ commerce-sdk.code-workspace
@@ -21,7 +21,7 @@
 			"name": "Integration Tests",
 			"path": "integrationTests"
 		}
-	
+
 	],
 
 	"settings": {}
