+++ integrationTests/package.json
@@ -8,9 +8,10 @@
     "depcheck": "depcheck",
     "lint": "eslint . --ext .ts,.js --quiet --fix",
     "pretest": "npm run lint && npm run depcheck",
-    "test:ts": "nyc mocha \"tests/*.test.ts\"",
-    "test:js": "nyc mocha \"tests/*.test.js\"",
+    "test:ts": "nyc mocha \"tests/**/*.test.ts\"",
+    "test:js": "nyc mocha \"tests/**/*.test.js\" --exclude \"tests/redis/**/*.test.js\"",
     "test": "npm run test:ts && npm run test:js",
+    "test:redis": "nyc mocha \"tests/redis/**/*.test.js\"",
     "redis": "npm run redis --prefix ../packages/core",
     "redis:shutdown": "npm run redis:shutdown --prefix ../packages/core",
     "redis:debug": "npm run redis:debug --prefix ../packages/core",
@@ -21,10 +22,10 @@
     "bundle:install": "npm run bundle:all && npm install",
     "bundle:clean:install": "npm run clean:all && npm run bundle:install",
     "test:integration": "npm run test",
-    "test:integration:redis": "npm run redis && npm run test; npm run redis:shutdown",
+    "test:integration:redis": "npm run redis && npm run test:redis; npm run redis:shutdown",
     "test:integration:redis:debug": "npm run redis:debug && npm run test; npm run redis:debug:shutdown",
-    "test:integration:bundle": "npm run bundle:install && npm run test",
-    "test:integration:bundle:redis": "npm run bundle:install && npm run redis && npm run test; npm run redis:shutdown",
+    "test:integration:bundle": "npm run bundle:install && npm run test && npm run test:redis",
+    "test:integration:bundle:redis": "npm run bundle:install && npm run redis && npm run test && npm run test:redis; npm run redis:shutdown",
     "test:integration:bundle:redis:debug": "npm run bundle:install && npm run redis:debug && npm run test; npm run redis:debug:shutdown",
     "clean": "rm -rf commerce-*.tgz || true",
     "clean:node": "rm -rf node_modules || true",
@@ -137,6 +138,7 @@
     "eslint-plugin-prettier": "^3.1.3",
     "esm": "^3.2.25",
     "mocha": "^7.1.2",
+    "nock": "^12.0.3",
     "nyc": "^15.0.1"
   }
 }
