+++ integrationTests/tests/cacheHelpers/delayedTests.test.helper.js
@@ -1,16 +1,20 @@
 /*
- * Copyright (c) 2019, salesforce.com, inc.
+ * Copyright (c) 2020, salesforce.com, inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 "use strict";
-import chai from "chai";
-import nock from "nock";
 
-import { _get } from "../../src/base/staticClient";
+const chai = require("chai");
+const nock = require("nock");
+const { StaticClient } = require("@commerce-apps/core");
 
-export default function() {
+/**
+ * Integration tests with 2 seconds delay to verify the cached content
+ * for Salesforce Commerce SDK cache manager interface.
+ */
+module.exports = function() {
   const expect = chai.expect;
 
   describe("base client etag based delayed conditional get tests", function() {
@@ -28,27 +32,27 @@ export default function() {
         );
 
       // make first request and get response from server
-      _get({
+      StaticClient.get({
         client: client,
         path: "/unmodified"
       }).then(data => {
         //ensure response body is correct from server
-        expect(data).to.eql({ mock: "data" });
+        expect(data).to.deep.equal({ mock: "data" });
         //ensure all http calls are done
         //expect(nock.isDone()).to.be.true;
 
         scope.get("/unmodified").reply(304, function() {
           //verify if sdk adds if-none-match header
-          expect(this.req.headers["if-none-match"][0]).to.eql("etag");
+          expect(this.req.headers["if-none-match"][0]).to.deep.equal("etag");
         });
 
         setTimeout(function() {
-          _get({
+          StaticClient.get({
             client: client,
             path: "/unmodified"
           }).then(data => {
             //ensure content is not empty and equals to the cached content
-            expect(data).to.eql({ mock: "data" });
+            expect(data).to.deep.equal({ mock: "data" });
             expect(nock.isDone()).to.be.true;
             done();
           });
@@ -67,36 +71,38 @@ export default function() {
         );
 
       // make first request and get response from server
-      _get({
+      StaticClient.get({
         client: client,
         path: "/modified"
       }).then(data => {
         //ensure response body is correct from server
-        expect(data).to.eql({ mock: "data" });
+        expect(data).to.deep.equal({ mock: "data" });
         //ensure all http calls are done
         expect(nock.isDone()).to.be.true;
 
         scope.get("/modified").reply(
           304,
           function() {
-            expect(this.req.headers["if-none-match"][0]).to.eql("etag");
+            expect(this.req.headers["if-none-match"][0]).to.deep.equal("etag");
           },
           { ETag: "new etag", "Cache-Control": "must-revalidate, max-age=1" }
         );
 
         setTimeout(function() {
-          _get({
+          StaticClient.get({
             client: client,
             path: "/modified"
           }).then(data => {
             //ensure content is not empty and equals to the cached content
-            expect(data).to.eql({ mock: "data" });
+            expect(data).to.deep.equal({ mock: "data" });
             expect(nock.isDone()).to.be.true;
 
             scope.get("/modified").reply(
               304,
               function() {
-                expect(this.req.headers["if-none-match"][0]).to.eql("new etag");
+                expect(this.req.headers["if-none-match"][0]).to.deep.equal(
+                  "new etag"
+                );
               },
               {
                 ETag: "new etag",
@@ -105,12 +111,12 @@ export default function() {
             );
 
             setTimeout(function() {
-              _get({
+              StaticClient.get({
                 client: client,
                 path: "/modified"
               }).then(data => {
                 //ensure content is not empty and equals to the cached content
-                expect(data).to.eql({ mock: "data" });
+                expect(data).to.deep.equal({ mock: "data" });
                 expect(nock.isDone()).to.be.true;
                 done();
               });
@@ -127,11 +133,7 @@ export default function() {
     it("sdk adds if-modified-since header and returns cached content on 304 response", function(done) {
       const client = this.client;
 
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const dateLastModified = new Date(Date.now() - 10000000).toUTCString();
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const nowTimestamp = Date.now();
       //set past expiry, since ttl is not set by default, this asset should be cached
       const dateExpires = new Date(nowTimestamp - 100000).toUTCString();
@@ -149,27 +151,27 @@ export default function() {
         );
 
       // make first request and get response from server
-      _get({
+      StaticClient.get({
         client: client,
         path: "/lastmodified"
       }).then(data => {
         //ensure response body is correct from server
-        expect(data).to.eql({ mock: "data" });
+        expect(data).to.deep.equal({ mock: "data" });
         //ensure all http calls are done
         expect(nock.isDone()).to.be.true;
 
         scope.get("/lastmodified").reply(304, function() {
-          expect(this.req.headers["if-modified-since"][0]).to.eql(
+          expect(this.req.headers["if-modified-since"][0]).to.deep.equal(
             dateLastModified
           );
         });
 
-        _get({
+        StaticClient.get({
           client: client,
           path: "/lastmodified"
         }).then(data => {
           //ensure content is not empty and equals to the cached content
-          expect(data).to.eql({ mock: "data" });
+          expect(data).to.deep.equal({ mock: "data" });
           expect(nock.isDone()).to.be.true;
           done();
         });
@@ -178,21 +180,10 @@ export default function() {
 
     it("sdk adds updated if-modified-since header on 304 response", function(done) {
       const client = this.client;
-
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const nowTimestamp = Date.now();
-
       const lastModified = new Date(nowTimestamp - 20000000).toUTCString();
-
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const newLastModified = new Date(nowTimestamp - 10000000).toUTCString();
 
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const dateExpires = new Date(nowTimestamp - 100000).toUTCString();
       const scope = nock("https://somewhere")
         .get("/not-modified")
@@ -207,19 +198,19 @@ export default function() {
         );
 
       // make first request and get response from server
-      _get({
+      StaticClient.get({
         client: client,
         path: "/not-modified"
       }).then(data => {
         //ensure response body is correct from server
-        expect(data).to.eql({ mock: "data" });
+        expect(data).to.deep.equal({ mock: "data" });
         //ensure all http calls are done
         expect(nock.isDone()).to.be.true;
 
         scope.get("/not-modified").reply(
           304,
           function() {
-            expect(this.req.headers["if-modified-since"][0]).to.eql(
+            expect(this.req.headers["if-modified-since"][0]).to.deep.equal(
               lastModified
             );
           },
@@ -230,24 +221,24 @@ export default function() {
           }
         );
 
-        _get({
+        StaticClient.get({
           client: client,
           path: "/not-modified"
         }).then(data => {
           //ensure content is not empty and equals to the cached content
-          expect(data).to.eql({ mock: "data" });
+          expect(data).to.deep.equal({ mock: "data" });
           expect(nock.isDone()).to.be.true;
           scope.get("/not-modified").reply(304, function() {
-            expect(this.req.headers["if-modified-since"][0]).to.eql(
+            expect(this.req.headers["if-modified-since"][0]).to.deep.equal(
               newLastModified
             );
           });
-          _get({
+          StaticClient.get({
             client: client,
             path: "/not-modified"
           }).then(data => {
             //ensure content is not empty and equals to the cached content
-            expect(data).to.eql({ mock: "data" });
+            expect(data).to.deep.equal({ mock: "data" });
             expect(nock.isDone()).to.be.true;
             done();
           });
@@ -258,12 +249,7 @@ export default function() {
     it("sdk adds if-modified-since header and returns modified content on 200 response", function(done) {
       const client = this.client;
       const nowTimestamp = Date.now();
-
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const dateLastModified = new Date(nowTimestamp - 10000000).toUTCString();
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       const dateExpires = new Date(nowTimestamp - 100000).toUTCString();
       const scope = nock("https://somewhere")
         .get("/lastmodified-since")
@@ -278,32 +264,32 @@ export default function() {
         );
 
       // make first request and get response from server
-      _get({
+      StaticClient.get({
         client: client,
         path: "/lastmodified-since"
       }).then(data => {
         //ensure response body is correct from server
-        expect(data).to.eql({ mock: "data" });
+        expect(data).to.deep.equal({ mock: "data" });
         //ensure all http calls are done
         expect(nock.isDone()).to.be.true;
 
         scope.get("/lastmodified-since").reply(200, function() {
-          expect(this.req.headers["if-modified-since"][0]).to.eql(
+          expect(this.req.headers["if-modified-since"][0]).to.deep.equal(
             dateLastModified
           );
           return { mock: "data_modified" };
         });
 
-        _get({
+        StaticClient.get({
           client: client,
           path: "/lastmodified-since"
         }).then(data => {
           //ensure content is not empty and equals to the modified content
-          expect(data).to.eql({ mock: "data_modified" });
+          expect(data).to.deep.equal({ mock: "data_modified" });
           expect(nock.isDone()).to.be.true;
           done();
         });
       });
     });
   });
-}
+};
