+++ integrationTests/tests/cacheHelpers/noCacheHeader.test.helper.js
@@ -1,16 +1,21 @@
 /*
- * Copyright (c) 2019, salesforce.com, inc.
+ * Copyright (c) 2020, salesforce.com, inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 "use strict";
-import chai from "chai";
-import nock from "nock";
 
-import { _get } from "../../src/base/staticClient";
+const chai = require("chai");
+const nock = require("nock");
 
-export default function() {
+const { StaticClient } = require("@commerce-apps/core");
+
+/**
+ * Basic tests that verifies cached content in absence of cache headers
+ * for Salesforce Commerce SDK cache manager interface.
+ */
+module.exports = function() {
   const expect = chai.expect;
   const RESPONSE_DATA = { mock: "data" };
   const RESPONSE_DATA_MODIFIED = { mock: "data_modified" };
@@ -24,20 +29,22 @@ export default function() {
         .get("/private")
         .reply(200, RESPONSE_DATA, { "Cache-Control": "private" });
 
-      return _get({ client: client, path: "/private" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
-        expect(nock.isDone()).to.be.true;
-        scope
-          .get("/private")
-          .reply(200, RESPONSE_DATA_MODIFIED, { "Cache-Control": "private" });
-        return _get({
-          client: client,
-          path: "/private"
-        }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+      return StaticClient.get({ client: client, path: "/private" }).then(
+        data => {
+          expect(data).to.deep.equal(RESPONSE_DATA);
           expect(nock.isDone()).to.be.true;
-        });
-      });
+          scope
+            .get("/private")
+            .reply(200, RESPONSE_DATA_MODIFIED, { "Cache-Control": "private" });
+          return StaticClient.get({
+            client: client,
+            path: "/private"
+          }).then(data => {
+            expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
+            expect(nock.isDone()).to.be.true;
+          });
+        }
+      );
     });
 
     it("asset not cached on response header no-cache", function() {
@@ -46,20 +53,22 @@ export default function() {
         .get("/no-cache")
         .reply(200, RESPONSE_DATA, { "Cache-Control": "no-cache" });
 
-      return _get({ client: client, path: "/no-cache" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
-        expect(nock.isDone()).to.be.true;
-        scope
-          .get("/no-cache")
-          .reply(200, RESPONSE_DATA_MODIFIED, { "Cache-Control": "no-cache" });
-        return _get({
-          client: client,
-          path: "/no-cache"
-        }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+      return StaticClient.get({ client: client, path: "/no-cache" }).then(
+        data => {
+          expect(data).to.deep.equal(RESPONSE_DATA);
           expect(nock.isDone()).to.be.true;
-        });
-      });
+          scope.get("/no-cache").reply(200, RESPONSE_DATA_MODIFIED, {
+            "Cache-Control": "no-cache"
+          });
+          return StaticClient.get({
+            client: client,
+            path: "/no-cache"
+          }).then(data => {
+            expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
+            expect(nock.isDone()).to.be.true;
+          });
+        }
+      );
     });
   });
 
@@ -72,20 +81,22 @@ export default function() {
         .get("/no-store")
         .reply(200, RESPONSE_DATA, { "Cache-Control": "no-store" });
 
-      return _get({ client: client, path: "/no-store" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
-        expect(nock.isDone()).to.be.true;
-        scope
-          .get("/no-store")
-          .reply(200, RESPONSE_DATA_MODIFIED, { "Cache-Control": "no-store" });
-        return _get({
-          client: client,
-          path: "/no-store"
-        }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+      return StaticClient.get({ client: client, path: "/no-store" }).then(
+        data => {
+          expect(data).to.deep.equal(RESPONSE_DATA);
           expect(nock.isDone()).to.be.true;
-        });
-      });
+          scope.get("/no-store").reply(200, RESPONSE_DATA_MODIFIED, {
+            "Cache-Control": "no-store"
+          });
+          return StaticClient.get({
+            client: client,
+            path: "/no-store"
+          }).then(data => {
+            expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
+            expect(nock.isDone()).to.be.true;
+          });
+        }
+      );
     });
   });
 
@@ -100,15 +111,18 @@ export default function() {
           Expires: "Wed, 18 Mar 2000 00:00:00 GMT"
         });
 
-      return _get({ client: client, path: "/invalid-expires" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+      return StaticClient.get({
+        client: client,
+        path: "/invalid-expires"
+      }).then(data => {
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope.get("/invalid-expires").reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/invalid-expires"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -122,18 +136,20 @@ export default function() {
           Expires: "Wed, 18 Mar 2050 00:00:00 GMT"
         });
 
-      return _get({ client: client, path: "/valid-expires" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
-        expect(nock.isDone()).to.be.true;
-        scope.get("/valid-expires").reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
-          client: client,
-          path: "/valid-expires"
-        }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA);
-          expect(nock.isDone()).to.be.false;
-        });
-      });
+      return StaticClient.get({ client: client, path: "/valid-expires" }).then(
+        data => {
+          expect(data).to.deep.equal(RESPONSE_DATA);
+          expect(nock.isDone()).to.be.true;
+          scope.get("/valid-expires").reply(200, RESPONSE_DATA_MODIFIED);
+          return StaticClient.get({
+            client: client,
+            path: "/valid-expires"
+          }).then(data => {
+            expect(data).to.deep.equal(RESPONSE_DATA);
+            expect(nock.isDone()).to.be.false;
+          });
+        }
+      );
     });
 
     it("asset not cached on response headers Expires and no-cache", function() {
@@ -145,20 +161,20 @@ export default function() {
           "Cache-Control": "no-cache"
         });
 
-      return _get({
+      return StaticClient.get({
         client: client,
         path: "/valid-expires-with-no-cache"
       }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope
           .get("/valid-expires-with-no-cache")
           .reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/valid-expires-with-no-cache"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -173,20 +189,20 @@ export default function() {
           "Cache-Control": "no-cache"
         });
 
-      return _get({
+      return StaticClient.get({
         client: client,
         path: "/valid-expires-with-no-store"
       }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope
           .get("/valid-expires-with-no-store")
           .reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/valid-expires-with-no-store"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -201,20 +217,20 @@ export default function() {
           "Cache-Control": "max-age=-1"
         });
 
-      return _get({
+      return StaticClient.get({
         client: client,
         path: "/valid-expires-with-invalid-max-age"
       }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope
           .get("/valid-expires-with-invalid-max-age")
           .reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/valid-expires-with-invalid-max-age"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -232,15 +248,18 @@ export default function() {
           "Cache-Control": "max-age=-1"
         });
 
-      return _get({ client: client, path: "/invalid-max-age" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+      return StaticClient.get({
+        client: client,
+        path: "/invalid-max-age"
+      }).then(data => {
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope.get("/invalid-max-age").reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/invalid-max-age"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -254,18 +273,20 @@ export default function() {
           "Cache-Control": "max-age=0"
         });
 
-      return _get({ client: client, path: "/max-age-zero" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
-        expect(nock.isDone()).to.be.true;
-        scope.get("/max-age-zero").reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
-          client: client,
-          path: "/max-age-zero"
-        }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+      return StaticClient.get({ client: client, path: "/max-age-zero" }).then(
+        data => {
+          expect(data).to.deep.equal(RESPONSE_DATA);
           expect(nock.isDone()).to.be.true;
-        });
-      });
+          scope.get("/max-age-zero").reply(200, RESPONSE_DATA_MODIFIED);
+          return StaticClient.get({
+            client: client,
+            path: "/max-age-zero"
+          }).then(data => {
+            expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
+            expect(nock.isDone()).to.be.true;
+          });
+        }
+      );
     });
 
     it("asset cached on response header max-age=10000000", function() {
@@ -276,15 +297,18 @@ export default function() {
           "Cache-Control": "max-age=10000000"
         });
 
-      return _get({ client: client, path: "/max-age-10000000" }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+      return StaticClient.get({
+        client: client,
+        path: "/max-age-10000000"
+      }).then(data => {
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope.get("/max-age-10000000").reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/max-age-10000000"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA);
+          expect(data).to.deep.equal(RESPONSE_DATA);
           expect(nock.isDone()).to.be.false;
         });
       });
@@ -298,20 +322,20 @@ export default function() {
           "Cache-Control": "max-age=10000000, no-cache "
         });
 
-      return _get({
+      return StaticClient.get({
         client: client,
         path: "/max-age-10000000-and-no-cache"
       }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope
           .get("/max-age-10000000-and-no-cache")
           .reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/max-age-10000000-and-no-cache"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
@@ -325,23 +349,23 @@ export default function() {
           "Cache-Control": "max-age=10000000, no-store"
         });
 
-      return _get({
+      return StaticClient.get({
         client: client,
         path: "/max-age-10000000-and-no-store"
       }).then(data => {
-        expect(data).to.eql(RESPONSE_DATA);
+        expect(data).to.deep.equal(RESPONSE_DATA);
         expect(nock.isDone()).to.be.true;
         scope
           .get("/max-age-10000000-and-no-store")
           .reply(200, RESPONSE_DATA_MODIFIED);
-        return _get({
+        return StaticClient.get({
           client: client,
           path: "/max-age-10000000-and-no-store"
         }).then(data => {
-          expect(data).to.eql(RESPONSE_DATA_MODIFIED);
+          expect(data).to.deep.equal(RESPONSE_DATA_MODIFIED);
           expect(nock.isDone()).to.be.true;
         });
       });
     });
   });
-}
+};
