+++ integrationTests/tests/redis/redisCachingDriver.test.js
@@ -1,27 +1,21 @@
 /*
- * Copyright (c) 2019, salesforce.com, inc.
+ * Copyright (c) 2020, salesforce.com, inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 "use strict";
-import chai from "chai";
-import chaiAsPromised from "chai-as-promised";
 
-import delayedTests from "../test/cache/delayedTests.tests";
-import cacheTests from "../test/cache/basic.tests";
-import etagTests from "../test/cache/etag.tests";
-import evictionTests from "../test/cache/eviction.tests";
-import multipleHeadersTests from "../test/cache/multipleHeaders.tests";
-import noCacheHeaderTests from "../test/cache/noCacheHeader.tests";
-
-import { BaseClient } from "../src/base/client";
-import { CacheManagerRedis } from "../src/base/cacheManagerRedis";
+const { BaseClient, CacheManagerRedis } = require("@commerce-apps/core");
+const cacheTests = require("../cacheHelpers/basic.test.helper");
+const etagTests = require("../cacheHelpers/etag.test.helper");
+const evictionTests = require("../cacheHelpers/eviction.test.helper");
+const multipleHeadersTests = require("../cacheHelpers/multipleHeaders.test.helper");
+const noCacheHeaderTests = require("../cacheHelpers/noCacheHeader.test.helper");
+const delayedTests = require("../cacheHelpers/delayedTests.test.helper");
 
 describe("Redis cache tests", function() {
   before(function() {
-    chai.should();
-    chai.use(chaiAsPromised);
     this.client = new BaseClient({
       baseUri: "https://somewhere",
       cacheManager: new CacheManagerRedis({
