+++ package.json
@@ -15,15 +15,14 @@
     "test:integration:bundle:redis": "lerna run test:integration:redis",
     "test:integration:bundle:redis:debug": "lerna run test:integration:redis:debug",
     "test:integration": "npm run test:integration:bundle --prefix ./integrationTests",
-    "test:integration:redis": "npm run test:integration:redis --prefix ./integrationTests",
-    "test:integration:redis:debug": "npm run test:integration:redis:debug --prefix ./integrationTests",
+    "test:integration:redis": "npm run test:integration:bundle:redis --prefix ./integrationTests",
+    "test:integration:redis:debug": "npm run test:integration:bundle:redis:debug --prefix ./integrationTests",
     "test:all": "npm run test && npm run test:integration",
-    "test:all:redis": "npm run test && npm run test:integration:bundle:redis",
-    "test:all:redis:debug": "npm run test && npm run test:integration:bundle:redis:debug",
+    "test:all:redis": "npm run test && npm run test:integration:redis",
+    "test:all:redis:debug": "npm run test && npm run test:integration:redis:debug",
     "clean": "lerna run clean && npm run clean --prefix ./integrationTests",
     "clean:node": "lerna run clean:node && npm run clean:node --prefix ./integrationTests",
     "clean:all": "npm run clean && npm run clean:node && npm run clean:all --prefix ./integrationTests",
-    "test:cache": "lerna run test:cache",
     "test:core": "lerna run --scope @commerce-apps/core test:debug",
     "test:generator": "lerna run --scope commerce-sdk test:debug",
     "publish": "lerna publish from-git --yes"
