+++ packages/core/package.json
@@ -11,7 +11,6 @@
     "build": "tsc",
     "test": "nyc mocha \"test/**/*.test.ts\" --reporter=xunit --reporter-options output=../../reports/core.xml",
     "test:debug": "nyc mocha \"test/**/*.test.ts\"",
-    "test:cache": "nyc --check-coverage=false mocha \"test_integration/**/*.ts\" --reporter=xunit --reporter-options output=../../reports/core-integration.xml --retries 2 --timeout 5000",
     "clean": "rm -rf package-lock.json && rm -rf dist",
     "clean:node": "rm -rf node_modules",
     "clean:all": "npm run clean && npm run clean:node",
